% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read files from Federal Accident
Reporting System
This is a function to read in a file
sourced from the Federal Accident
Reporting System.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string
that specifies the path and name of
the file sourced from FARS

 @return This function returns a data
 frame that has the dataset imported
 from the source file read in from FARS

 @examples fars_read(fn)

 @export}
}
\description{
Read files from Federal Accident
Reporting System
This is a function to read in a file
sourced from the Federal Accident
Reporting System.
}
